+++ sei-cosmos/x/staking/types/params_test.go
@@ -1,6 +1,7 @@
 package types_test
 
 import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -20,4 +21,18 @@ func TestParamsEqual(t *testing.T) {
 
 	ok = p1.Equal(p2)
 	require.False(t, ok)
+}
+
+func Test_validateParams(t *testing.T) {
+	params := types.DefaultParams()
+
+	// default params have no error
+	require.NoError(t, params.Validate())
+
+	// validate mincommision
+	params.MinCommissionRate = sdk.NewDec(-1)
+	require.Error(t, params.Validate())
+
+	params.MinCommissionRate = sdk.NewDec(2)
+	require.Error(t, params.Validate())
 }